LayoutTests/http/tests/dom/window-open-about-uppercase-blank-and-access-document-expected.txt

CONSOLE MESSAGE: line 1: Injected script running.
PASS newWindow.document is defined.
